Rework ThinBox to use middle-pointer
* Changes:

- Implement ThinBox using a middle-pointer.
- Remove the now unnecessary AlignedMetaData trait.

* Motivations:

Obtaining the size/alignment of a type just from its data could paint
ourselves in a corner with regard to the design of fat-pointers.

As it turns out that ThinBox::deref() can be implemented in a more
efficient fashion using a middle-pointer -- less instructions, fixed
offsets -- then do so and remove the need for AlignedMetaData.